langchain-tool/pyproject.toml
[project]
name = "langchain-tool"
version = "0.1.0"
description = "Langchain tool learning" 
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv==1.0.0",
    "anthropic>=0.40.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "wikipedia>=1.4.0",
    "numpy<2.0",
    "langgraph>=0.1.0",
]
